feat(loja): reutilizar formulario de inscricao nos eventos
