+++ app/loja/eventos/page.tsx
@@ -4,6 +4,7 @@
 import Image from "next/image";
 import { useEffect, useState } from "react";
 import { formatDate } from "@/utils/formatDate";
+import InscricaoForm from "../components/InscricaoForm";
 
 interface Evento {
   id: string;
@@ -22,14 +23,14 @@ export default function EventosPage() {
     const tenantId = localStorage.getItem("tenant_id");
     fetch(`/api/eventos?tenant=${tenantId ?? ""}`)
       .then((r) => r.json())
-      .then((data) => Array.isArray(data) ? setEventos(data) : setEventos([]))
+      .then((data) => (Array.isArray(data) ? setEventos(data) : setEventos([])))
       .catch((err) => {
         console.error("Erro ao carregar eventos:", err);
       });
   }, []);
 
   return (
-    <main className="px-4 py-10 md:px-16">
+    <main className="px-4 py-10 md:px-16 space-y-10">
       <h1 className="text-3xl font-bold text-center mb-10">Eventos UMADEUS</h1>
       <div className="grid gap-8 md:grid-cols-2 lg:grid-cols-3">
         {eventos.map((evento) => (
@@ -64,6 +65,8 @@ export default function EventosPage() {
           </div>
         ))}
       </div>
+      <InscricaoForm />
     </main>
   );
 }
+
